doc it
